+++ tests/pest/Feature/BlitzVariableTest.php
@@ -6,20 +6,32 @@
 
 use putyourlightson\blitz\variables\BlitzVariable;
 
-test('Include cached tag does not contain encoded slashes in params', function() {
+test('Include cached tag does not contain unencoded slashes in params', function() {
     $variable = new BlitzVariable();
     $tagString = (string)$variable->includeCached('test');
+    preg_match('/_includes\?(.*)/', $tagString, $match);
 
-    expect($tagString)
+    expect($match[1])
         ->not()
-        ->toContain('%2F');
+        ->toContain('/');
 });
 
-test('Fetch URI tag does not contain encoded slashes in params', function() {
+test('Include cached tag does not contain path param', function() {
+    $variable = new BlitzVariable();
+    $tagString = (string)$variable->includeCached('test');
+    preg_match('/\?(.*)/', $tagString, $match);
+
+    expect($match[1])
+        ->not()
+        ->toContain(Craft::$app->getConfig()->getGeneral()->pathParam . '=');
+});
+
+test('Fetch URI tag does not contain unencoded slashes in params', function() {
     $variable = new BlitzVariable();
     $tagString = (string)$variable->fetchUri('test', ['action' => 'x/y/z']);
+    preg_match('/blitz-params="(.*?)"/', $tagString, $match);
 
-    expect($tagString)
+    expect($match[1])
         ->not()
-        ->toContain('%2F');
+        ->toContain('/');
 });